Implementing stateless plugin system

+++ packages/data_manager.py
@@ -20,14 +20,15 @@ class DataManager:
         #Change this to 'REST_AI_Server' during deployment
         self.db = self.client["test"]
 
-        self.collections_list = ["robot_status", "logs", "infos"]
+        self.collections_list = ["robot_status", "logs", "infos", "plugins"]
 
         #Inserts collections with base entry if necessary
         for collection in self.collections_list:
             if collection not in self.db.list_collection_names():
                 new_col = self.db[collection]
                 if collection == "robot_status":
-                    dict = {"token" : "", 
+                    dict = {"token" : "",
+                            "confirmation" : False,
                             "command" : "", 
                             "parameters": {
                                 "type": "",
@@ -46,6 +47,9 @@ class DataManager:
                     base_entry = new_col.insert_one(dict)
                 elif collection == "infos":
                     dict = {"token" : "", "msg": ""}
+                    base_entry = new_col.insert_one(dict)
+                elif collection == "plugins":
+                    dict = {"token": "", "plugin_id": ""}
                     base_entry = new_col.insert_one(dict)
 
         print("Database is set-up!")
@@ -97,22 +101,64 @@ class DataManager:
 
     #Saves the dictionary in the specified collection
     def save_data(self, collection, dict:dict):
-        chosen_collection = self.db[collection]
+        if collection in self.collections_list:
+            chosen_collection = self.db[collection]
 
-        #Regarding robot_status: In order to avoid complex queries and ensure the return of only the latest position, prior entries will be deleted
-        if collection == "robot_status":
-            token = dict["token"]
-            doc = self.query_data("robot_status", {"token": token})
-            if len(doc) != 0:
-                chosen_collection.delete_many({"token": token})
-        entry = chosen_collection.insert_one(dict)
+            #Regarding robot_status: In order to avoid complex queries and 
+            #ensure the return of only the latest position, prior entries will be deleted
+            #Currently not checking correct formats -> separate method
+            if collection == "robot_status":
+                token = dict["token"]
+                doc = self.query_data("robot_status", {"token": token})
+                #Checks for previous entries, deletes them
+                if len(doc) != 0:
+                    chosen_collection.delete_many({"token": token})
+            chosen_collection.insert_one(dict)
     
     #Returns a list of the requested documents after taking in a query dictionary
+    #So far only used for database testing
     def query_data(self, collection, query_dict):
         if collection in self.collections_list:
             chosen_collection = self.db[collection]
             #Returns a cursor object and turns it into a list of dictionaries
             document = list(chosen_collection.find(query_dict))
             return document
+    
+    #Sets the confirmation status to true
+    #After receiving a post /newcommand request
+    def confirm_robot_command(self, token):
+        collection = self.db["robot_status"]
+        query =  {"token" : token}
+        #Setting new values
+        new_values = {"$set" : {"confirmation" : True}}
+        #Updating database
+        collection.update_one(query, new_values)
+
+#Testing
+# dm = DataManager()
+# doc = {"token" : "1234678",
+#                             "confirmation" : False,
+#                             "command" : "", 
+#                             "parameters": {
+#                                 "type": "",
+#                                 "frame": {
+#                                     "X": 0, 
+#                                     "Y": 0, 
+#                                     "Z": 0, 
+#                                     "A": 0, 
+#                                     "B": 0, 
+#                                     "Z": 0}
+#                             }
+#                     }
+# dm.save_data("robot_status", doc)
+# print(dm.query_data("robot_status", doc))
+# dm.confirm_robot_command("1234678")
+# print(dm.query_data("robot_status", {"token" : "1234678"}))
+# dm.stop_mongodb()
+
+
+        
+
+
